api-backend/requirements.txt
aniso8601==9.0.1
anyio==3.7.1
bcrypt==4.0.1
bleach==6.0.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cors==1.0.1
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
flask-mailman==0.3.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
future==0.18.3
gevent==23.7.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mkdocs-material-extensions==1.1.1
passlib==1.7.4
pycparser==2.21
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1
tldextract==3.4.4
typing_extensions==4.7.1
urllib3==2.0.4
webencodings==0.5.1
Werkzeug==2.3.6
WTForms==3.0.1
zope.event==5.0
zope.interface==6.0
